refactor: remove either wrapper from jsonToBlob
