+++ packages/hoppscotch-app/helpers/functional/json.ts
@@ -15,3 +15,8 @@ export const safeParseJSON = (str: string): O.Option<object> =>
  * @returns If string is a JSON string
  */
 export const isJSON = flow(safeParseJSON, O.isSome)
+
+export const jsonToBlob = (content: object) =>
+  new Blob([JSON.stringify(content)], {
+    type: "application/json",
+  })
