+++ packages/hoppscotch-app/helpers/import-export/export/hopp.ts
@@ -2,9 +2,9 @@ import { HoppRESTRequest } from "@hoppscotch/data"
 import { pipe } from "fp-ts/function"
 import * as TE from "fp-ts/TaskEither"
 import { HoppExporter } from "."
-import { jsonToBlob } from "~/helpers/utils/export"
+import { jsonToBlob } from "~/helpers/functional/json"
 
-const exporter: HoppExporter<HoppRESTRequest, "CANNOT_MAKE_BLOB"> = (content) =>
-  pipe(content, jsonToBlob, TE.fromEither)
+const exporter: HoppExporter<HoppRESTRequest, never> = (content) =>
+  pipe(content, jsonToBlob, TE.right)
 
 export default exporter
